LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/api/basic/integrity.sub.any-expected.txt

CONSOLE MESSAGE: Failed integrity metadata check. http://web-platform.test:8800/fetch/api/resources/top.txt. Failed integrity metadata check. Content length: 3, Expected content length: 0, Expected metadata: sha256-dKUcPOn/AlUjWIwcHeHNqYXPlvyGiq+2dWOdFcE+24I=
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/top.txt due to access control checks.
CONSOLE MESSAGE: Failed integrity metadata check. http://web-platform.test:8800/fetch/api/resources/top.txt. Failed integrity metadata check. Content length: 3, Expected content length: 0, Expected metadata: sha512-oUceBRNxPxnY60g/VtPCj2syT4wo4EZh2CgYdWy9veW8+OsReTXoh7dizMGZafvx9+QhMS39L/gIkxnPIn41Zg== sha384-MgZYnnAzPM/MjhqfOIMfQK5qcFvGZsGLzx4Phd7/A8fHTqqLqXqKo8cNzY3xEPTL
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/top.txt due to access control checks.
CONSOLE MESSAGE: Failed integrity metadata check. http://web-platform.test:8800/fetch/api/resources/top.txt. Failed integrity metadata check. Content length: 3, Expected content length: 0, Expected metadata: sha256-dKUcPOn/AlUjWIwcHeHNqYXPlvyGiq+2dWOdFcE+24I= sha512-oUceBRNxPxnY60g/VtPCj2syT4wo4EZh2CgYdWy9veW8+OsReTXoh7dizMGZafvx9+QhMS39L/gIkxnPIn41Zg==
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/top.txt due to access control checks.
CONSOLE MESSAGE: Failed integrity metadata check. http://web-platform.test:8801/fetch/api/resources/top.txt?pipe=header(Access-Control-Allow-Origin,*). Failed integrity metadata check. Content length: 3, Expected content length: 0, Expected metadata: sha512-oUceBRNxPxnY60g/VtPCj2syT4wo4EZh2CgYdWy9veW8+OsReTXoh7dizMGZafvx9+QhMS39L/gIkxnPIn41Zg==
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8801/fetch/api/resources/top.txt?pipe=header(Access-Control-Allow-Origin,*) due to access control checks.
CONSOLE MESSAGE: Failed integrity metadata check. https://web-platform.test:9443/fetch/api/resources/top.txt. Failed integrity metadata check. Content length: 3, Expected content length: 0, Expected metadata: sha512-D6yns0qxG0E7+TwkevZ4Jt5t7Iy3ugmAajG/dlf6Pado1JqTyneKXICDiqFIkLMRExgtvg8PlxbKTkYfRejSOg==
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/fetch/api/resources/top.txt due to access control checks.

PASS Empty string integrity
PASS SHA-256 integrity
PASS SHA-384 integrity
PASS SHA-512 integrity
PASS SHA-512 integrity with missing padding
PASS SHA-512 integrity base64url encoded
PASS SHA-512 integrity base64url encoded with missing padding
PASS Invalid integrity
PASS Multiple integrities: valid stronger than invalid
PASS Multiple integrities: invalid stronger than valid
PASS Multiple integrities: invalid as strong as valid
PASS Multiple integrities: both are valid
PASS Multiple integrities: both are invalid
PASS CORS empty integrity
PASS CORS SHA-512 integrity
PASS CORS invalid integrity
PASS Empty string integrity for opaque response
PASS SHA-* integrity for opaque response
